Added the capability to embed Python instead of extending it.  This creates one, big monolithic executable, 'pycitcoms', which contains all CitcomS code and an embedded Python interpreter.  Embedding works-around MPI portability issues.  (See issue15 for details.) Embed by default.
Configure with '--disable-embedding' to build the old way with ".so"
extension modules.  This is *NOT* guaranteed to work unless you have a
shared MPI library.  The ugly hacks which made it work in the past
(even with a static "libmpich.a", for example) are no longer
supported.

Merged "Full" and "Regional" into a single library to cut down on
redundancy, both in the source code and in the amount of object code
which is produced.  (I performed this merge so that I would only need
a single 'pycitcoms' intepreter, instead of two.)