Revised everything having to do with runtime configuration: launcher.xxx, scheduler.xxx, and job.xxx.
Lowered the prominence/importance given to .pml files...
and resisted my temptation to delete all reference to them :-)

(I think I'm pretty much done with the v2.1 manual at this point.
Reading the manual makes me tired, because it reminds me of how
much work remains to be done on the code itself.  Ease-of-use
could be substantially improved...)