Fixed some bugs, some of which were introduced during the C -> C++ transition.
Using a pidXXXX.cfg file as input should "work" now, although it still
doesn't reproduce the original simulation, since at least one
parameter is missing ('steps').  I maintain that dumping the
configuration is really the job of the framework (Pyre).
Unfortunately, Pyre's current capabilities in this area are buggy in
the case of '.pml' files, and non-existent in the case of '.cfg'
files.

Note: 'nproc_surf' is now a real inventory item, instead of a fake
one.  It looks like it doesn't belong in the inventory at all, though.
(I left it there to avoid breakage.)