Fix for infinite regress when builddir == srcdir.  !@#$%*
The .pth/sitedir/sys.path/PYTHONPATH aspect of eggs is really
irritating.  I hope setuptools becomes a part of the standard
Python distro, as the Python core developers are likely to
accept it uncritically and unmodified (as they are wont to do,
apparently -- judging by all the other crap that is in the
standard library).  God willing, this will further accelerate
the death of the language.