Compute the velocity in Cartesian coord on the fly.
The previous code computed and stored the Cartesian velocity in E->temp in
v_from_vector() and accessed E->temp in velo_from_element. This is bug prone.
E->temp is a temporary array and should be used for storage.

Also, in previous code, if E->sphere.cap[m].V is modified, one has to recompute
E->temp. Otherwise, the Cartesian velocity will be incorrect.