Adding the ability to combine optional fields to the post-processing scripts.
autocombine.py will read the parameter 'output_optional' from the cfg file and
combine the optional fields (can be pressure, stress, comp_nd, or a combination
of the three) into a file. The file is named similar to the cap file, with
'cap' replaced by 'opt'. A .general file is also created, which describes the
content, shape, and structure of the opt file.