Implemented writing and reading (binary) checkpoints.
The checkpoint file is called [datafile].chkpt.[rank] and is outputed with an 
interval of controller.checkpointFrequency. Before writing the checkpoint file,
the previous checkpoint file (if exists) is renamed to 
[datafile].chkpt.[rank].bak.

To read back the checkpoint file, set solver.ic.restart=on. The old checkpoint
file, [datafile_old].chkpt.[rank] will be read in. The timestep information is
stored in the checkpoint file. There is no need to specify solution_cycles_init,
as the old way of restarting.

The old way of restarting (read in temperature from the velo files) will still
be available through solver.ic.tic_method=-1. (To do.)

Several things to do: clean up the old way of restarting tracers/composition,
documentation, does mat_control need checkpoints?