Compute E->surf_det (top surface area, ie. surface jacobian, of an element) on a unit sphere.
Use E->surf_det, instead of the surface jacobian returned by get_global_1d_shape_fn(), to expand spherical harmonics. 

The surface jacobian returned by get_global_1d_shape_fn() is defined at the top radius of the current mesh. When nprocz>1, the top radius is not the same as the outer radius of the sphere for processors. As a result, the spherical harmonics expansion gave incorrect result when nprocz>1 in the previous revisions.