Added parameters 'amending_outflow' and 'exchange_pressure' to help the convergence of esolver
When 'amending_outflow' is set to true, the imposed velocity BC will be
amended slightly to be divergence-free. (The divergence, e.g. outflow, is
caused by the combination of solver inaccuracy and interpolation inaccuracy.)

When 'exchange_pressure' is set to true, the initial pressure (at element
level) of the embedded solver is taken from the pressure solution of the
containing solver.