This version seeks to merge the tracer and strain-tracking routines coded up as part of
Fuchs, L. and Becker, T. W.: Role of strain-dependent weakening memory
on the style of mantle convection and plate boundary
stability. Geophys. J. Int., 218, 601-618, 2019.

and the in review 2022 work on global scales. I verified that the code
compiles, and expect these changes to not affect any default behavior,
but still need to do more testing.